+++ client/src/GalleryPage/mapObjects/Balloon.tsx
@@ -1,0 +1,22 @@
+import { animated, Interpolation } from "@react-spring/three";
+import { MeshDistortMaterial } from "@react-three/drei";
+import { COLORS } from "../../@types/colors";
+
+const AnimatedMeshDistortMaterial = animated(MeshDistortMaterial);
+
+interface BallonProps {
+  position: [x: number, y: number, z: number];
+  positionY: Interpolation<number, number>;
+  color: Interpolation<number, COLORS>;
+}
+
+export default function Balloon({ position, positionY, color }: BallonProps) {
+  return (
+    <animated.mesh position={position} position-y={positionY}>
+      <sphereGeometry args={[1.5, 64, 32]} />
+      {/* eslint-disable-next-line @typescript-eslint/ban-ts-comment */}
+      {/* @ts-ignore */}
+      <AnimatedMeshDistortMaterial speed={5} distort={0.3} color={color} />
+    </animated.mesh>
+  );
+}
